fix: partial refactor for setup_first-second_relay_node
